feat(app): Allow for lazy registration of app components (#15)
* WIP: add lazy module instantiation

* feat(app): allow for lazy loading of services

* test(app): add integration test for webpack/browserify for lazy instantiation

* fix(app): fix issue where auth listeners weren't being patched properly

* docs(app): adding doc explaining Firebase Auth specific code in firebase_app.ts

* fix(app): revert code refactor to _getService function

* test(app): add tests and fix issue with multiple instances of a service

There was an issue where we weren't properly limiting mulitple service instances. Added a test and
refactored the code to support that use case

* refactor(app): remove unneeded ternary

* fix(app): fix issue where instanceIdentifier wasn't being passed

Fixed an issue where instance identifiers weren't being passed to their creation factories. Added
some tests to validate that we don't accidentally remove that in the future.