Make Timestamp a part of the public Firestore API. (#544)
* Make Timestamp a part of the public Firestore API.

* Accept Timestamps as a parameter where Dates are currently accepted.

* Add an option to settings to control how timestamp fields are returned.

This is a transitional state; eventually, Timestamp will live in a new
separate package (common-types/).

The behavior to return timestamp fields as dates is now legacy. It's on
by default to allow for a transitional period. The new behavior to
return Timestamp objects instead is opt-in, but will become default in
the future. The advantage Timestamps have is higher precision than
native browser Date. If the new behavior is not enabled, a warning is
logged on startup urging users to upgrade.